split protocol
